Fix uint64 uptime_ms failing with NANOPB_WITHOUT_64BIT
CONFIG_ZMK_STUDIO implies CONFIG_NANOPB_WITHOUT_64BIT, which redefines
pb_uint64_t = uint32_t. pb_enc_varint checks field->data_size against
sizeof(pb_uint64_t), so a uint64_t field (data_size=8) never matches any
branch and returns "invalid data_size", causing pb_get_encoded_size to fail.

Fix: change RuntimeStatus.uptime_ms from uint64 to uint32. The value wraps
every ~49 days which is acceptable for a diagnostic uptime counter.

Document the constraint in AGENTS.md so future proto fields don't repeat it.

+++ src/studio/device_info_handler.c
@@ -144,7 +144,7 @@ static int handle_get_device_info(const zmk_device_info_GetDeviceInfoRequest *re
 
     /* Runtime status */
     result.has_runtime = true;
-    result.runtime.uptime_ms = (uint64_t)k_uptime_get();
+    result.runtime.uptime_ms = (uint32_t)k_uptime_get();
 
     resp->which_response_type = zmk_device_info_Response_device_info_tag;
     resp->response_type.device_info = result;
